+++ okitweb/static/okit/model/js/artefacts/customer_premise_equipment.js
@@ -28,23 +28,13 @@ class CustomerPremiseEquipment extends OkitArtifact {
     ** Clone Functionality
      */
     clone() {
-        return new CustomerPremiseEquipment(this, this.getOkitJson());
+        return new CustomerPremiseEquipment(JSON.clone(this), this.getOkitJson());
     }
 
 
     /*
     ** Delete Processing
      */
-    deleteChildren() {
-        // Remove IPSec references
-        for (let ipsec of this.getOkitJson().ipsec_connections) {
-            for (let i=0; i < ipsec.customer_premise_equipments.length; i++) {
-                if (ipsec.customer_premise_equipments[i] === this.id) {
-                    ipsec.customer_premise_equipments.splice(i, 1);
-                }
-            }
-        }
-    }
 
 
     getNamePrefix() {