+++ okitweb/static/okit/view/designer/js/artefacts/mysql_database_system.js
@@ -14,6 +14,8 @@ class MySQLDatabaseSystemView extends OkitDesignerArtefactView {
 
     get parent_id() {return this.artefact.subnet_id;}
     get parent() {return this.getJsonView().getSubnet(this.parent_id);}
+    // Direct Subnet Access
+    set subnet_id(id) {this.artefact.subnet_id = id;}
 
     /*
      ** SVG Processing