+++ okitweb/static/okit/view/js/okit_view.js
@@ -160,10 +160,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteAutonomousDatabase(id='') {
-        this.okitjson.deleteAutonomousDatabase(id);
-        this.update();
-    }
     loadAutonomousDatabases(autonomous_databases) {
         for (const artefact of autonomous_databases) {
             this.autonomous_databases.push(new AutonomousDatabaseView(new AutonomousDatabase(artefact, this.okitjson), this));
@@ -181,9 +177,7 @@ class OkitJsonView {
         this.block_storage_volumes.push(volume ? new BlockStorageVolumeView(volume, this) : new BlockStorageVolumeView(this.okitjson.newBlockStorageVolume(), this));
         return this.block_storage_volumes[this.block_storage_volumes.length - 1];
     }
-    getBlockStorageVolumes() {
-        return this.block_storage_volumes;
-    }
+    getBlockStorageVolumes() {return this.block_storage_volumes;}
     getBlockStorageVolume(id='') {
         for (let artefact of this.getBlockStorageVolumes()) {
             if (artefact.id === id) {
@@ -191,10 +185,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteBlockStorageVolume(id='') {
-        this.okitjson.deleteBlockStorageVolume(id);
-        this.update();
     }
     loadBlockStorageVolumes(block_storage_volumes) {
         for (const artefact of block_storage_volumes) {
@@ -222,13 +212,19 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteCompartment(id='') {
-        this.okitjson.deleteCompartment(id);
-        this.update();
-    }
     loadCompartments(compartments) {
         for (const artefact of compartments) {
             this.compartments.push(new CompartmentView(new Compartment(artefact, this.okitjson), this));
+        }
+    }
+    loadCompartmentsSelect(select_id, empty_option=false) {
+        $(jqId(select_id)).empty();
+        const compartment_select = $(jqId(select_id));
+        if (empty_option) {
+            compartment_select.append($('<option>').attr('value', '').text(''));
+        }
+        for (let compartment of this.getCompartments()) {
+            compartment_select.append($('<option>').attr('value', compartment.id).text(compartment.display_name));
         }
     }
 
@@ -253,10 +249,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteCustomerPremiseEquipment(id='') {
-        this.okitjson.deleteCustomerPremiseEquipment(id);
-        this.update();
     }
     loadCustomerPremiseEquipments(fast_connects) {
         for (const artefact of fast_connects) {
@@ -287,10 +279,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteDatabaseSystem(id='') {
-        this.okitjson.deleteDatabaseSystem(id);
-        this.update();
-    }
     loadDatabaseSystems(database_systems) {
         for (const artefact of database_systems) {
             this.database_systems.push(new DatabaseSystemView(new DatabaseSystem(artefact, this.okitjson), this));
@@ -320,10 +308,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteDynamicRoutingGateway(id='') {
-        this.okitjson.deleteDynamicRoutingGateway(id);
-        this.update();
-    }
     loadDynamicRoutingGateways(dynamic_routing_gateways) {
         for (const artefact of dynamic_routing_gateways) {
             this.dynamic_routing_gateways.push(new DynamicRoutingGatewayView(new DynamicRoutingGateway(artefact, this.okitjson), this));
@@ -351,10 +335,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteFastConnect(id='') {
-        this.okitjson.deleteFastConnect(id);
-        this.update();
     }
     loadFastConnects(fast_connects) {
         for (const artefact of fast_connects) {
@@ -384,10 +364,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteFileStorageSystem(id='') {
-        this.okitjson.deleteFileStorageSystem(id);
-        this.update();
     }
     loadFileStorageSystems(file_storage_systems) {
         for (const artefact of file_storage_systems) {
@@ -422,10 +398,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteInstance(id='') {
-        this.okitjson.deleteInstance(id);
-        this.update();
-    }
     loadInstances(instances) {
         for (const artefact of instances) {
             this.instances.push(new InstanceView(new Instance(artefact, this.okitjson), this));
@@ -454,10 +426,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteInstancePool(id='') {
-        this.okitjson.deleteInstancePool(id);
-        this.update();
     }
     loadInstancePools(instance_pools) {
         for (const artefact of instance_pools) {
@@ -499,10 +467,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteInternetGateway(id='') {
-        this.okitjson.deleteInternetGateway(id);
-        this.update();
-    }
     loadInternetGateways(internet_gateways) {
         for (const artefact of internet_gateways) {
             this.internet_gateways.push(new InternetGatewayView(new InternetGateway(artefact, this.okitjson), this));
@@ -530,10 +494,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteIPSecConnection(id='') {
-        this.okitjson.deleteIPSecConnection(id);
-        this.update();
     }
     loadIPSecConnections(fast_connects) {
         for (const artefact of fast_connects) {
@@ -564,10 +524,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteLoadBalancer(id='') {
-        this.okitjson.deleteLoadBalancer(id);
-        this.update();
-    }
     loadLoadBalancers(load_balancers) {
         for (const artefact of load_balancers) {
             this.load_balancers.push(new LoadBalancerView(new LoadBalancer(artefact, this.okitjson), this));
@@ -597,10 +553,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteLocalPeeringGateway(id='') {
-        this.okitjson.deleteLocalPeeringGateway(id);
-        this.update();
-    }
     loadLocalPeeringGateways(local_peering_gateways) {
         for (const artefact of local_peering_gateways) {
             this.local_peering_gateways.push(new LocalPeeringGatewayView(new LocalPeeringGateway(artefact, this.okitjson), this));
@@ -629,10 +581,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteMySQLDatabaseSystem(id='') {
-        this.okitjson.deleteMySQLDatabaseSystem(id);
-        this.update();
     }
     loadMySQLDatabaseSystems(database_systems) {
         for (const artefact of database_systems) {
@@ -674,10 +622,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteNATGateway(id='') {
-        this.okitjson.deleteNATGateway(id);
-        this.update();
-    }
     loadNATGateways(nat_gateways) {
         for (const artefact of nat_gateways) {
             this.nat_gateways.push(new NATGatewayView(new NATGateway(artefact, this.okitjson), this));
@@ -707,13 +651,35 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteNetworkSecurityGroup(id='') {
-        this.okitjson.deleteNetworkSecurityGroup(id);
-        this.update();
-    }
     loadNetworkSecurityGroups(network_security_groups) {
         for (const artefact of network_security_groups) {
             this.network_security_groups.push(new NetworkSecurityGroupView(new NetworkSecurityGroup(artefact, this.okitjson), this));
+        }
+    }
+    loadNetworkSecurityGroupsSelect(select_id, empty_option=false) {
+        $(jqId(select_id)).empty();
+        const nsg_select = $(jqId(select_id));
+        if (empty_option) {
+            nsg_select.append($('<option>').attr('value', '').text(''));
+        }
+        for (let nsg of this.getNetworkSecurityGroups()) {
+            nsg_select.append($('<option>').attr('value', nsg.id).text(nsg.display_name));
+        }
+    }
+    loadNetworkSecurityGroupsMultiSelect(select_id, vcn_id) {
+        $(jqId(select_id)).empty();
+        const multi_select = d3.select(d3Id(select_id));
+        for (let nsg of this.getNetworkSecurityGroups()) {
+            if (nsg.vcn_id === vcn_id || !vcn_id) {
+                const div = multi_select.append('div');
+                div.append('input')
+                    .attr('type', 'checkbox')
+                    .attr('id', safeId(nsg.id))
+                    .attr('value', nsg.id);
+                div.append('label')
+                    .attr('for', safeId(nsg.id))
+                    .text(nsg.display_name);
+            }
         }
     }
 
@@ -738,10 +704,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteObjectStorageBucket(id='') {
-        this.okitjson.deleteObjectStorageBucket(id);
-        this.update();
     }
     loadObjectStorageBuckets(object_storage_buckets) {
         for (const artefact of object_storage_buckets) {
@@ -772,10 +734,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteOkeCluster(id='') {
-        this.okitjson.deleteOkeCluster(id);
-        this.update();
-    }
     loadOkeClusters(oke_clusters) {
         for (const artefact of oke_clusters) {
             this.oke_clusters.push(new OkeClusterView(new OkeCluster(artefact, this.okitjson), this));
@@ -803,10 +761,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteRemotePeeringConnection(id='') {
-        this.okitjson.deleteRemotePeeringConnection(id);
-        this.update();
     }
     loadRemotePeeringConnections(fast_connects) {
         for (const artefact of fast_connects) {
@@ -837,10 +791,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteRouteTable(id='') {
-        this.okitjson.deleteRouteTable(id);
-        this.update();
-    }
     loadRouteTables(route_tables) {
         for (const artefact of route_tables) {
             this.route_tables.push(new RouteTableView(new RouteTable(artefact, this.okitjson), this));
@@ -869,10 +819,6 @@ class OkitJsonView {
             }
         }
         return undefined;
-    }
-    deleteSecurityList(id='') {
-        this.okitjson.deleteSecurityList(id);
-        this.update();
     }
     loadSecurityLists(security_lists) {
         for (const artefact of security_lists) {
@@ -910,10 +856,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteServiceGateway(id='') {
-        this.okitjson.deleteServiceGateway(id);
-        this.update();
-    }
     loadServiceGateways(service_gateways) {
         for (const artefact of service_gateways) {
             this.service_gateways.push(new ServiceGatewayView(new ServiceGateway(artefact, this.okitjson), this));
@@ -933,9 +875,7 @@ class OkitJsonView {
         this.subnets.push(subnet ? new SubnetView(subnet, this) : new SubnetView(this.okitjson.newSubnet(), this));
         return this.subnets[this.subnets.length - 1];
     }
-    getSubnets() {
-        return this.subnets;
-    }
+    getSubnets() {return this.subnets;}
     getSubnet(id='') {
         for (let artefact of this.getSubnets()) {
             if (artefact.id === id) {
@@ -944,13 +884,22 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteSubnet(id='') {
-        this.okitjson.deleteSubnet(id);
-        this.update();
-    }
     loadSubnets(subnets) {
         for (const artefact of subnets) {
             this.subnets.push(new SubnetView(new Subnet(artefact, this.okitjson), this));
+        }
+    }
+    loadSubnetsSelect(select_id, empty_option=false) {
+        $(jqId(select_id)).empty();
+        const subnet_select = $(jqId(select_id));
+        if (empty_option) {
+            subnet_select.append($('<option>').attr('value', '').text(''));
+        }
+        for (let subnet of this.getSubnets()) {
+            const compartment = this.getCompartment(subnet.compartment_id);
+            const vcn = this.getVirtualCloudNetwork(subnet.vcn_id);
+            const display_name = `${compartment.display_name}/${vcn.display_name}/${subnet.display_name}`;
+            subnet_select.append($('<option>').attr('value', subnet.id).text(display_name));
         }
     }
 
@@ -987,13 +936,21 @@ class OkitJsonView {
     getVcn(id='') {
         return this.getVirtualCloudNetwork(id);
     }
-    deleteVirtualCloudNetwork(id='') {
-        this.okitjson.deleteVirtualCloudNetwork(id);
-        this.update();
-    }
     loadVirtualCloudNetworks(virtual_cloud_networks) {
         for (const artefact of virtual_cloud_networks) {
             this.virtual_cloud_networks.push(new VirtualCloudNetworkView(new VirtualCloudNetwork(artefact, this.okitjson), this));
+        }
+    }
+    loadVirtualCloudNetworksSelect(select_id, empty_option=false) {
+        $(jqId(select_id)).empty();
+        const vcn_select = $(jqId(select_id));
+        if (empty_option) {
+            vcn_select.append($('<option>').attr('value', '').text(''));
+        }
+        for (let vcn of this.getVirtualCloudNetworks()) {
+            const compartment = this.getCompartment(vcn.compartment_id);
+            const display_name = `${compartment.display_name}/${vcn.display_name}`;
+            vcn_select.append($('<option>').attr('value', vcn.id).text(display_name));
         }
     }
 
@@ -1011,7 +968,6 @@ class OkitJsonView {
         }
         return undefined;
     }
-    deleteVirtualNetworkInterface(id='') {}
 
     // Fragment
     dropFragmentView(target) {
@@ -1021,35 +977,50 @@ class OkitJsonView {
     newFragment(target) {
         return new FragmentView(this.okitjson.newFragment(target), this);
     }
-
 }
 
 /*
 ** Simple Artefact View Class for all artefacts that are not Containers
  */
 class OkitArtefactView {
+    static cut_copy_paste = {resource: undefined, paste_count: 0, is_cut: false};
+
     constructor(artefact=null, json_view) {
         this.artefact = artefact;
         this.collapsed = false;
         this._recalculate_dimensions = true;
-        this.getJsonView = function() {return json_view};
         // Raise Artefact Elements to View Class
         if (this.artefact) {
-            for (let key in this.artefact) {
-                Object.defineProperty(this, key, { get: function() {return this.artefact[key];} });
-            }
+            Object.entries(this.artefact).forEach(
+                ([key, value]) => {
+                    if (!(value instanceof Function)) {
+                        Object.defineProperty(this, key, {
+                            get: function () {
+                                return this.artefact[key];
+                            }
+                        });
+                    }
+                }
+            );
         }
+        this.getJsonView = function() {return json_view};
+        this.getOkitJson = function() {return json_view.getOkitJson()};
     }
 
     // -- Reference
+    get resource_name() {return this.getArtifactReference();}
     get json_view() {return this.getJsonView();}
-    get okit_json() {return this.json_view.getOkitJson();}
+    get okit_json() {return this.getJsonView().getOkitJson();}
+    get list_name() {return `${this.resource_name.toLowerCase().split(' ').join('_')}s`;}
+    get json_model_list() {return this.okit_json[this.list_name];}
+    get json_view_list() {return this.json_view[this.list_name];}
     //get id() {return this.artefact ? this.artefact.id : '';}
     get artefact_id() {return this.artefact ? this.artefact.id : '';}
     get attached() {return false;}
     get compartment_id() {return this.artefact ? this.artefact.compartment_id : '';}
     get parent_id() {return null;}
     get parent() {return null;}
+    get children() {return [];}
     get display_name() {return this.artefact ? this.artefact.display_name : '';}
     get definition() {return this.artefact ? this.artefact.definition : '';}
     get is_collapsed() {return this.parent ? this.collapsed || this.parent.is_collapsed : this.collapsed;}
@@ -1115,7 +1086,7 @@ class OkitArtefactView {
     }
     // --- Dimensions
     get recalculate_dimensions() {return this._recalculate_dimensions;}
-    set recalculate_dimensions(recalculate) {this._recalculate_dimensions = true; this.parent ? this.parent.recalculate_dimensions = true : recalculate = false;}
+    set recalculate_dimensions(recalculate) {this._recalculate_dimensions = true;this.parent ? this.parent.recalculate_dimensions = true : recalculate = false;}
     get width_multiplier() {return this.show_label ? okitSettings.show_label === 'name' ? 1.5 : 2 : 1;}
     get height_multiplier() {return this.show_label ?  1.5 : 1;}
     get icon_dimensions() {return {width: this.icon_width, height: this.icon_height};}
@@ -1291,10 +1262,54 @@ class OkitArtefactView {
     get summary_tooltip() {return this.display_name;}
     // ---- Connectors
     get top_bottom_connectors_preferred() {return true;}
+    // ---- Okit View Functions
+    get new_function() {return `new${this.getArtifactReference().split(' ').join('')}`}
+    get cloneable() {return true;}
+    get moveable() {return true;}
+    get pasteableOrig() {return this.json_view.copied_artefact ? this.json_view.copied_artefact.getDropTargets().includes(this.getArtifactReference()) : false;}
+    get pasteable() {return OkitArtefactView.cut_copy_paste.resource ? OkitArtefactView.cut_copy_paste.resource.getDropTargets().includes(this.getArtifactReference()) : false;}
+    get pasteableNew() {
+        return OkitArtefactView.cut_copy_paste.resource ? OkitArtefactView.cut_copy_paste.resource.getDropTargets().includes(this.getArtifactReference()) : false;
+    }
 
     getArtefact() {return this.artefact;}
 
-    getOkitJson() {return this.okit_json;}
+    static new(artefact, json_view) {return new this(artefact, json_view);}
+
+    cut() {OkitArtefactView.cut_copy_paste.resource = this; OkitArtefactView.cut_copy_paste.paste_count = 0; this.json_view.is_cut = true; this.deleteSvg();}
+
+    copy() {OkitArtefactView.cut_copy_paste.resource = this; OkitArtefactView.cut_copy_paste.paste_count = 0; this.json_view.is_cut = false;}
+
+    paste(drop_target) {
+        const clone = OkitArtefactView.cut_copy_paste.resource.artefact.clone();
+        if (!OkitArtefactView.cut_copy_paste.is_cut) clone.display_name += 'Copy';
+        if (OkitArtefactView.cut_copy_paste.paste_count > 0) {clone.display_name += `-${OkitArtefactView.cut_copy_paste.paste_count}`;}
+        OkitArtefactView.cut_copy_paste.paste_count += 1;
+        clone.id = clone.okit_id;
+        drop_target.updateCloneIds(clone);
+        this.json_model_list.push(clone);
+        return clone;
+    }
+
+    clone() {
+        const clone = this.artefact.clone();
+        if (!this.json_view.is_cut) clone.display_name += 'Copy';
+        clone.id = clone.okit_id;
+        this.json_model_list.push(clone);
+        return clone;
+    }
+
+    delete() {
+        for (let i = 0; i < this.json_model_list.length; i++) {
+            if (this.json_model_list[i].id === this.id) {
+                this.json_model_list[i].delete();
+                this.json_model_list.splice(i, 1);
+                break;
+            }
+        }
+        // Remove SVG Element
+        if ($(jqId(this.svg_id)).length) {$(jqId(this.svg_id)).remove()}
+    }
 
     draw() {
         if ((!this.parent || !this.parent.is_collapsed) && (!okitSettings.hide_attached || !this.attached)) {
@@ -1421,6 +1436,7 @@ class OkitArtefactView {
         svg.on("click", function() {
             self.loadSlidePanels();
             d3.event.stopPropagation();
+            $(jqId("context-menu")).addClass("hidden");
         });
     }
 
@@ -1451,7 +1467,90 @@ class OkitArtefactView {
     }
 
     addContextMenu(svg) {
-        svg.on("contextmenu", handleContextMenu);
+        const self = this;
+        svg.on("contextmenu", function () {
+            d3.event.preventDefault();
+            d3.event.stopPropagation();
+            const canvas_position = $(jqId("canvas-div")).offset();
+            const position = {top: d3.event.pageY - canvas_position.top, left: d3.event.pageX - 5};
+            $(jqId("context-menu")).empty();
+            $(jqId("context-menu")).css(position);
+            const contextmenu = d3.select(d3Id("context-menu"));
+            contextmenu.on('mouseenter', function () {
+                    $(jqId("context-menu")).removeClass("hidden");
+                })
+                .on('mouseleave', function () {
+                    $(jqId("context-menu")).addClass("hidden");
+                });
+
+            contextmenu.append('label')
+                .attr('class', 'okit-context-menu-title')
+                .text(self.display_name)
+            const ul = contextmenu.append('ul')
+                .attr('class', 'okit-context-menu-list');
+            if (self.compartment_id) {
+                // Delete
+                ul.append('li').append('a')
+                    .attr('class', 'parent-item')
+                    .attr('href', 'javascript:void(0)')
+                    .text('Delete')
+                    .on('click', function () {
+                        self.delete();
+                        self.json_view.update(self.okit_json);
+                        $(jqId("context-menu")).addClass("hidden");
+                    });
+                // Cut
+                if (self.moveable) {
+                    ul.append('li').append('a')
+                        .attr('class', 'parent-item')
+                        .attr('href', 'javascript:void(0)')
+                        .text('Cut')
+                        .on('click', function () {
+                            OkitArtefactView.cut_copy_paste = {paste_count: 0, is_cut: true};
+                            self.cut();
+                            //self.json_view.update(self.okit_json);
+                            $(jqId("context-menu")).addClass("hidden");
+                        });
+                }
+                // Clone
+                if (self.cloneable) {
+                    ul.append('li').append('a')
+                        .attr('class', 'parent-item')
+                        .attr('href', 'javascript:void(0)')
+                        .text('Clone')
+                        .on('click', function () {
+                            self.clone();
+                            self.json_view.update(self.okit_json);
+                            $(jqId("context-menu")).addClass("hidden");
+                        });
+                }
+                // Copy
+                ul.append('li').append('a')
+                    .attr('class', 'parent-item')
+                    .attr('href', 'javascript:void(0)')
+                    .text('Copy')
+                    .on('click', function () {
+                        OkitArtefactView.cut_copy_paste = {paste_count: 0, is_cut: false};
+                        self.copy();
+                        $(jqId("context-menu")).addClass("hidden");
+                    });
+                $(jqId("context-menu")).removeClass("hidden");
+            }
+            // Paste
+            if (self.pasteable) {
+                ul.append('li').append('a')
+                    .attr('class', 'parent-item')
+                    .attr('href', 'javascript:void(0)')
+                    .text(`Paste ${OkitArtefactView.cut_copy_paste.resource.getArtifactReference()} ${OkitArtefactView.cut_copy_paste.resource.display_name}`)
+                    .on('click', function () {
+                        OkitArtefactView.cut_copy_paste.resource.paste(self);
+                        if (OkitArtefactView.cut_copy_paste.is_cut) OkitArtefactView.cut_copy_paste.resource.delete();
+                        self.json_view.update(self.okit_json);
+                        $(jqId("context-menu")).addClass("hidden");
+                    });
+                $(jqId("context-menu")).removeClass("hidden");
+            }
+        });
     }
 
     addCustomAttributes(svg) {
@@ -1696,9 +1795,9 @@ class OkitArtefactView {
         return coord_str;
     }
 
-    delete() {
+    deleteSvg() {
         // Remove SVG Element
-        if ($(jqId(this.id + "-svg")).length) {$(jqId(this.id + "-svg")).remove()}
+        if ($(jqId(this.svg_id)).length) {$(jqId(this.svg_id)).remove()}
     }
 
     /*
@@ -2042,12 +2141,22 @@ class OkitArtefactView {
         if (this.hasLeftChildren()) {
             offset.dx += Math.round(positional_adjustments.padding.x + positional_adjustments.spacing.x);
         }
+        let dx_adjustment = 0;
+        if (this.hasTopChildren()) {
+            let dimensions = this.getTopChildrenMaxDimensions();
+            dx_adjustment = Math.max(dimensions.width, dx_adjustment);
+        }
         if (this.hasContainerChildren()) {
             let dimensions = this.getContainerChildrenMaxDimensions();
-            offset.dx += dimensions.width;
-            offset.dx += positional_adjustments.spacing.x;
-            offset.dx += positional_adjustments.padding.x;
+            dx_adjustment = Math.max(dimensions.width, dx_adjustment);
         }
+        if (this.hasBottomChildren()) {
+            let dimensions = this.getBottomChildrenMaxDimensions();
+            dx_adjustment = Math.max(dimensions.width, dx_adjustment);
+        }
+        offset.dx += dx_adjustment;
+        offset.dx += positional_adjustments.spacing.x;
+        offset.dx += positional_adjustments.padding.x;
         return offset;
     }
 
@@ -2067,6 +2176,20 @@ class OkitArtefactView {
             }
         }
         return children;
+    }
+
+    getRightEdgeChildrenMaxDimensions() {
+        let max_dimensions = {height: 0, width: 0};
+        for (let group of this.getRightEdgeArtifacts()) {
+            for(let artefact of this.json_view[this.artefact.artefactToElement(group)]) {
+                if (artefact.parent_id === this.id) {
+                    let dimension = artefact.dimensions;
+                    max_dimensions.height += Math.round(dimension.height + positional_adjustments.spacing.y);
+                    max_dimensions.width = Math.max(max_dimensions.width, dimension.width);
+                }
+            }
+        }
+        return max_dimensions;
     }
 
     getFirstRightEdgeChildOffset() {
@@ -2204,36 +2327,36 @@ class OkitContainerArtefactView extends OkitArtefactView {
             let offset = {dx: 0, dy: 0};
             // Process Top Edge Artifacts
             offset = this.getFirstTopEdgeChildOffset();
-            let top_edge_dimensions = this.getTopEdgeChildrenMaxDimensions();
+            const top_edge_dimensions = this.getTopEdgeChildrenMaxDimensions();
             dimensions.width = Math.max(dimensions.width, top_edge_dimensions.width + offset.dx - padding.dx);
             dimensions.height = Math.max(dimensions.height, top_edge_dimensions.height);
             // Process Top Artifacts
             offset = this.getFirstTopChildOffset();
-            let top_dimensions = this.getTopChildrenMaxDimensions();
+            const top_dimensions = this.getTopChildrenMaxDimensions();
             dimensions.width = Math.max(dimensions.width, top_dimensions.width);
             dimensions.height += top_dimensions.height;
             // Process Container Artifacts
             offset = this.getFirstContainerChildOffset();
-            let container_dimensions = this.getContainerChildrenMaxDimensions();
+            const container_dimensions = this.getContainerChildrenMaxDimensions();
             dimensions.width = Math.max(dimensions.width, container_dimensions.width);
             dimensions.height += container_dimensions.height;
             // Process Bottom Artifacts
             offset = this.getFirstBottomChildOffset();
-            let bottom_dimensions = this.getBottomChildrenMaxDimensions();
+            const bottom_dimensions = this.getBottomChildrenMaxDimensions();
             dimensions.width = Math.max(dimensions.width, bottom_dimensions.width);
             dimensions.height += bottom_dimensions.height;
             // Process Bottom Edge Artifacts
             offset = this.getFirstBottomEdgeChildOffset();
-            let bottom_edge_dimensions = this.getBottomEdgeChildrenMaxDimensions();
+            const bottom_edge_dimensions = this.getBottomEdgeChildrenMaxDimensions();
             dimensions.width = Math.max(dimensions.width, bottom_edge_dimensions.width);
             dimensions.height = Math.max(dimensions.height, bottom_edge_dimensions.height);
             // Process Left Edge Artifacts
             // Process Left Artifacts
-            let left_dimensions = this.getLeftChildrenMaxDimensions();
+            const left_dimensions = this.getLeftChildrenMaxDimensions();
             dimensions.width += left_dimensions.width;
             dimensions.height = Math.max(dimensions.height, left_dimensions.height);
             // Process Right Artifacts
-            let right_dimensions = this.getRightChildrenMaxDimensions();
+            const right_dimensions = this.getRightChildrenMaxDimensions();
             dimensions.width += right_dimensions.width;
             dimensions.height = Math.max(dimensions.height, right_dimensions.height);
             if (this.hasRightChildren()) {
@@ -2241,6 +2364,13 @@ class OkitContainerArtefactView extends OkitArtefactView {
                 dimensions.width += positional_adjustments.padding.x;
             }
             // Process Right Edge Artifacts
+            const right_edge_dimensions = this.getRightEdgeChildrenMaxDimensions();
+            dimensions.width += right_dimensions.width;
+            dimensions.height = Math.max(dimensions.height, right_edge_dimensions.height);
+            if (this.hasRightEdgeChildren()) {
+                dimensions.width += positional_adjustments.spacing.x;
+                dimensions.width += positional_adjustments.padding.x;
+            }
             // Add Padding
             dimensions.width += padding.dx * 2;
             dimensions.height += padding.dy * 2;
@@ -2271,16 +2401,37 @@ class OkitContainerArtefactView extends OkitArtefactView {
     get show_info() {return this.collapsed ? super.show_info : true;}
     // ----- Label
     get show_label() {return this.collapsed ? super.show_label : false;}
+    // ---- Okit View Functions
 
-    /*
-    ** SVG Functions
-     */
+    paste(drop_target) {
+        const clone = super.paste(drop_target);
+        this.cloneChildren(clone);
+        return clone;
+    }
+
+    updateCloneIds(clone) {
+        if (this.getArtifactReference() === Subnet.getArtifactReference()) {
+            clone.subnet_id = this.id;
+            clone.compartment_id = this.compartment_id;
+        } else if (this.getArtifactReference() === VirtualCloudNetwork.getArtifactReference()) {
+            clone.vcn_id = this.id;
+            clone.compartment_id = this.compartment_id;
+        } else {
+            clone.compartment_id = this.id;
+        }
+        return clone;
+    }
+
+/*
+** SVG Functions
+ */
     drawIcon(svg) {
         const icon = super.drawIcon(svg);
         // Add Click Event to toggle collapsed
         const self = this;
         icon.on("click", function() {
             self.collapsed = !self.collapsed;
+            self.recalculate_dimensions = true;
             self.getJsonView().draw();
         });
     }